Update patch set 2
Patch Set 2:

nebkat, that's specifically what the persist value I set in c1-common is for. Either you didn't merge that or you unset/reset it since installing. Also, nightlies aren't always expected to be upgraded with 100% success every time, if you were to do a clean install with this patch (which is how I tested it on SGS1), it would work exactly like planned.

Patch-set: 2