Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Not ready yet - as the commit message says, I think we should really turn this off on the variants of the GS2 that don't have NFC, which is most of them. I have a GT-I9100P (from Orange UK) which this works on, and I think the Korean "I9100" (not GT-) has it as well, but I don't know whether the CM GT-I9100 builds even work on that.

I don't really know how best to vary this feature. If it can be done at install time or boot time based on what model the radio reports, I think that would be ideal. Failing that, should it be a separate build, "galaxys2nfc"?

Patch-set: 1