Update patch set 1
Patch Set 1:

On Gingerbread, we could get NFC (but without Wallet) by adding a permissions XML and (if on an I777 base) some smali hacks in Preferences to remove the disable logic.

If we added the NFC bits from the I777 stock firmware and a permissions XML in the process of porting an I9100 firmware, we got everything but NDEF.  NDEF would crap out due to differences in STOP_APP_SWITCHES permissions logic - I777 firmwares were SignatureOrSystem, I9100 firmwares had this set to Signature.

https://github.com/Entropy512/hellraiser_i777/commit/7a6a0245be873e71462c7693472ac83510b06cb6

I tried firmware from I777 Gingerbread and from the SHW-M250K, never tried pulling it from one of the Nexus series.

Patch-set: 1