Update patch set 1
Patch Set 1:

To say that the improvement is astonishing would be an understatement.

Without this patch I can hardly get a lock even while on open space, and I never get accuracy less than 10 meters.

With this patch, I have an instant lock even indoors, the phone detects more satellites it can use (5-6 indoors, more than 12 on a balcony in my house), and the accuracy is almost always 5m.

Patch-set: 1