Update patch set 1
Patch Set 1:

I would have to agree with Andrew here.
One way to fix this might be to overload (create a wrapper) of the snensors HAL and read the light sensor values from the sysfs interface.
Or fix the orientation/magnetometer, but I see that as a bigger / more difficult task.

Creating an wrapper is time consuming, but atleast doable without too much fuss.

Patch-set: 1