Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

Originally I thought this was redundant as the second mic is already enabled by default.

It turns out that Samsung seems to be swapping the meanings of this control.

dualmic_enabled=false - Noise cancellation mic is ON
dualmic_enabled=true - Noise cancellation mic is OFF

Eventually this should be exposed in the phone UI instead of being hardcoded.

Patch-set: 1