Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

While I was hoping for the core sensor conversion factors fix to get more testing, since it's merged, let's go ahead with this one.

It seems good other than a 3-4 degree offset in compass between my I777 (opensource) and Note (Sammy HAL) - but this may just be device-to-device variation of a sensor that's always a bit prone to offsets and variation.

Patch-set: 1